+++ src/nemos/_hess.py
